Use SimplifyNodeOptions in foundry_simplify_node
